# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for Polleg that allows unibo students to answer exam exercises directly on the cartabinaria website",
        "title": "Polleg API",
        "contact": {
            "name": "Gabriele Genovese",
            "email": "gabriele.genovese2@studio.unibo.it"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/moderation/ban": {
            "get": {
                "description": "Get all banned users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "Get all banned users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.BannedUser"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ApiError"
                        }
                    }
                }
            },
            "post": {
                "description": "Ban or unban a user given its username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "Ban or unban a user",
                "parameters": [
                    {
                        "description": "Ban or unban a user",
                        "name": "banUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BanUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ApiError"
                        }
                    }
                }
            }
        },
        "/moderation/report/{id}": {
            "post": {
                "description": "Report an answer given its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "Report an answer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Answer id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Report cause",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ApiError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a report given its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "Delete a report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Answer id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ApiError"
                        }
                    }
                }
            }
        },
        "/moderation/reports": {
            "get": {
                "description": "Get all reports",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderation"
                ],
                "summary": "Get all reports",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Report"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.BanUserRequest": {
            "type": "object",
            "properties": {
                "ban": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.BannedUser": {
            "type": "object",
            "properties": {
                "banned_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user_avatar_url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "api.Report": {
            "type": "object",
            "properties": {
                "answer_id": {
                    "type": "integer"
                },
                "cause": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_avatar_url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.ReportRequest": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                }
            }
        },
...
        "httputil.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}